Fixes some translations in slate block picker
